Tests/LibWeb/Text/expected/ShadowDOM/css-hover-shadow-dom.txt

hovered bg: rgb(0, 128, 0)
